Update phone field type to allow longer numbers

+++ migrations/1-create-table-users.js
@@ -23,7 +23,7 @@ module.exports = {
         allowNull: false,
       },
       phone: {
-        type: Sequelize.INTEGER(15),
+        type: Sequelize.INTEGER(20),
         allowNull: false,
       },
       address: {
